P3A: refactor shields stats.
