+++ browser/extensions/api/brave_shields_api.cc
@@ -58,18 +58,6 @@ ExtensionFunction::ResponseAction BraveShieldsAllowScriptsOnceFunction::Run() {
   return RespondNow(NoArguments());
 }
 
-void RecordShieldsToggled() {
-  PrefService* local_state = g_browser_process->local_state();
-  ::brave_shields::RecordShieldsUsageP3A(::brave_shields::kShutOffShields,
-                                         local_state);
-}
-
-void RecordShieldsSettingChanged() {
-  PrefService* local_state = g_browser_process->local_state();
-  ::brave_shields::RecordShieldsUsageP3A(
-      ::brave_shields::kChangedPerSiteShields, local_state);
-}
-
 ExtensionFunction::ResponseAction
 BraveShieldsSetBraveShieldsEnabledFunction::Run() {
   std::unique_ptr<brave_shields::SetBraveShieldsEnabled::Params> params(
@@ -84,8 +72,6 @@ BraveShieldsSetBraveShieldsEnabledFunction::Run() {
 
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetBraveShieldsEnabled(profile, params->enabled, url);
-
-  RecordShieldsToggled();
 
   return RespondNow(NoArguments());
 }
@@ -128,8 +114,6 @@ ExtensionFunction::ResponseAction BraveShieldsSetAdControlTypeFunction::Run() {
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetAdControlType(profile, control_type, url);
 
-  RecordShieldsSettingChanged();
-
   return RespondNow(NoArguments());
 }
 
@@ -170,8 +154,6 @@ BraveShieldsSetCookieControlTypeFunction::Run() {
 
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetCookieControlType(profile, control_type, url);
-
-  RecordShieldsSettingChanged();
 
   return RespondNow(NoArguments());
 }
@@ -215,8 +197,6 @@ BraveShieldsSetFingerprintingControlTypeFunction::Run() {
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetFingerprintingControlType(profile, control_type, url);
 
-  RecordShieldsSettingChanged();
-
   return RespondNow(NoArguments());
 }
 
@@ -253,8 +233,6 @@ BraveShieldsSetHTTPSEverywhereEnabledFunction::Run() {
 
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetHTTPSEverywhereEnabled(profile, params->enabled, url);
-
-  RecordShieldsSettingChanged();
 
   return RespondNow(NoArguments());
 }
@@ -297,8 +275,6 @@ BraveShieldsSetNoScriptControlTypeFunction::Run() {
 
   Profile* profile = Profile::FromBrowserContext(browser_context());
   ::brave_shields::SetNoScriptControlType(profile, control_type, url);
-
-  RecordShieldsSettingChanged();
 
   return RespondNow(NoArguments());
 }
